+++ GitTrends.Shared/Models/IssueConnection.cs
@@ -6,13 +6,13 @@ namespace GitTrends.Shared
 {
     public class IssuesConnection
     {
-        public IssuesConnection(int totalCount, IEnumerable<Issue>? nodes) =>
+        public IssuesConnection(long totalCount, IEnumerable<Issue>? nodes) =>
             (IssuesCount, IssueList) = (totalCount, nodes?.ToList() ?? Enumerable.Empty<Issue>().ToList());
 
         [JsonProperty("nodes")]
         public List<Issue> IssueList { get; }
 
         [JsonProperty("totalCount")]
-        public int IssuesCount { get; }
+        public long IssuesCount { get; }
     }
 }
